Get the useAccount hook working
