+++ packages/hooks/src/useAccount.test.tsx
@@ -3,7 +3,7 @@ import {
   // act
 } from '@testing-library/react-hooks';
 
-import { useAccount } from '.';
+import { useAccount } from './useAccount';
 
 const INPUT_ENTROPY_ARRAY = [
   21,
@@ -38,45 +38,6 @@ test('does nothing without randomBytes', () => {
   );
 
   expect(result.current[0]).toEqual('');
-  expect(result.current[1]).toEqual(false);
-});
-
-test('handles rerendering with NO change in randomBytes input', () => {
-  const { result, rerender } = renderHook(() =>
-    useAccount({
-      randomBytes: new Uint8Array(),
-      randomBytesGenerated: false,
-    })
-  );
-
-  rerender();
-
-  expect(result.current[0]).toEqual('');
-  expect(result.current[1]).toEqual(false);
-});
-
-test('handles rerendering with a change in randomBytes', () => {
-  let randomBytes = new Uint8Array();
-  let randomBytesGenerated = false;
-
-  const { result, rerender } = renderHook(() =>
-    useAccount({
-      randomBytes: randomBytes,
-      randomBytesGenerated: randomBytesGenerated,
-    })
-  );
-
-  randomBytes = new Uint8Array(INPUT_ENTROPY_ARRAY);
-  randomBytesGenerated = true;
-
-  // TODO: Figure out how to get to rerender the component with a new input
-  // for the useAccount hook
-  rerender();
-
-  expect(result.current[0]).toEqual(
-    '0x80F8f3629b58b0b2873c6424cDe17540F645df16'
-  );
-  expect(result.current[1]).toEqual(true);
 });
 
 test('does something with randomBytes', () => {
@@ -90,5 +51,41 @@ test('does something with randomBytes', () => {
   expect(result.current[0]).toEqual(
     '0x80F8f3629b58b0b2873c6424cDe17540F645df16'
   );
-  expect(result.current[1]).toEqual(true);
+});
+
+test('handles rerendering with NO change in randomBytes input', () => {
+  const { result, rerender } = renderHook(() =>
+    useAccount({
+      randomBytes: new Uint8Array(),
+      randomBytesGenerated: false,
+    })
+  );
+
+  rerender();
+
+  expect(result.current[0]).toEqual('');
+});
+
+test('handles rerendering with a change in randomBytes', () => {
+  let randomBytes = new Uint8Array();
+  let randomBytesGenerated = false;
+
+  const { result, rerender } = renderHook(() =>
+    useAccount({
+      randomBytes: randomBytes,
+      randomBytesGenerated: randomBytesGenerated,
+    })
+  );
+  // TODO: Ensure there is no race condition here
+  // since the useEffect on first render triggers an async operation
+  // We could await the state update that happens internally in the hook
+  // when that async effect completes
+
+  randomBytes = new Uint8Array(INPUT_ENTROPY_ARRAY);
+  randomBytesGenerated = true;
+  rerender();
+
+  expect(result.current[0]).toEqual(
+    '0x80F8f3629b58b0b2873c6424cDe17540F645df16'
+  );
 });
